HTTPClient v0.1.0 HTTPClient.Adapter behaviour

An HTTP client behaviour definition.

Adapters that implement the HTTP client behaviour must support the following list of functions:

request/5 - Issues an HTTP request with the given method to the given url

get/3 - issues GET request to the given url

post/4 - issues POST request to the given url

put/4 - issues PUT request to the given url

patch/4 - issues PATCH request to the given url

delete/3 - issues DELETE request to the given url

All functions must return response defined in response/0.

Types

body()

Specs

body() ::
  HTTPClient.Adapters.Finch.body() | HTTPClient.Adapters.HTTPoison.body()

headers()

Specs

headers() ::
  HTTPClient.Adapters.Finch.headers() | HTTPClient.Adapters.HTTPoison.headers()

method()

Specs

method() ::
  HTTPClient.Adapters.Finch.method() | HTTPClient.Adapters.HTTPoison.method()

options()

Specs

options() ::
  HTTPClient.Adapters.Finch.options() | HTTPClient.Adapters.HTTPoison.options()

response()

Specs

response() :: {:ok, HTTPClient.Response.t()} | {:error, HTTPClient.Error.t()}

url()

Specs

url() :: HTTPClient.Adapters.Finch.url() | HTTPClient.Adapters.HTTPoison.url()

Functions

set(opts)

Sets HTTP client implementation adapter based on the use options.

Using options

:adapter - Implementation of adapter to use. The default value is :httpoison.

Callbacks

delete(url, headers, options)

Specs

delete(url(), headers(), options()) :: response()

get(url, headers, options)

Specs

get(url(), headers(), options()) :: response()

patch(url, body, headers, options)

Specs

patch(url(), body(), headers(), options()) :: response()

post(url, body, headers, options)

Specs

post(url(), body(), headers(), options()) :: response()

put(url, body, headers, options)

Specs

put(url(), body(), headers(), options()) :: response()

request(method, url, body, headers, options)

Specs

request(method(), url(), body(), headers(), options()) :: response()